Updating taikungoclient dependency (#246)
* Bumping taikungoclient and fixing build

* Updating command tree

* Bumping latest dependency

* Fixing build

* Bumping taikungoclient to latest

* Adding missing fields to some of the resources

* Fixing minor bugs

* Fixing pagination issues

* Updating tests

+++ cmd/cloudcredential/azure/list/list.go
@@ -2,6 +2,7 @@ package list
 
 import (
 	"context"
+
 	"github.com/itera-io/taikun-cli/api"
 	"github.com/itera-io/taikun-cli/cmd/cmdutils"
 	"github.com/itera-io/taikun-cli/config"
@@ -117,7 +118,7 @@ func ListCloudCredentialsAzure(opts *ListOptions) (credentials []interface{}, er
 			break
 		}
 
-		if count == data.GetTotalCount() {
+		if int64(count) == data.GetTotalCount() { // casting is safe, extending maximum bounds
 			break
 		}
