+++ cmd/project/list/list.go
@@ -2,6 +2,7 @@ package list
 
 import (
 	"context"
+
 	"github.com/itera-io/taikun-cli/api"
 	"github.com/itera-io/taikun-cli/cmd/cmdutils"
 	"github.com/itera-io/taikun-cli/config"
@@ -139,7 +140,7 @@ func listRun(opts *ListOptions) (err error) {
 			break
 		}
 
-		if projectsCount == data.GetTotalCount() {
+		if int64(projectsCount) == data.GetTotalCount() { // casting is safe, extending maximum bounds
 			break
 		}
