+++ cmd/slackconfiguration/list/list.go
@@ -2,6 +2,7 @@ package list
 
 import (
 	"context"
+
 	"github.com/itera-io/taikun-cli/api"
 	"github.com/itera-io/taikun-cli/cmd/cmdutils"
 	"github.com/itera-io/taikun-cli/config"
@@ -100,7 +101,7 @@ func listRun(opts *ListOptions) (err error) {
 			break
 		}
 
-		if count == data.GetTotalCount() {
+		if int64(count) == data.GetTotalCount() { // casting is safe, extending maximum bounds
 			break
 		}
